requirements.txt
apache-airflow-providers-google == 10.3.0
soda-core-bigquery==3.0.45
astronomer-cosmos[dbt-bigquery]
protobuf
